pyproject.toml
[project]
name = "contract-review-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "nltk>=3.9.1",
    "openai>=1.75.0",
    "pdfplumber>=0.11.6",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.13.3",
    "ruff>=0.11.6",
    "streamlit>=1.44.1",
]
